CommonRouter作成 同一処理をまとめました。クラスメソッドで良いのか悪いのか分かりませんが一旦これで進めます。
