+++ RssReader/Router/CommonRouter.swift
@@ -1,0 +1,49 @@
+//
+//  CommonRouter.swift
+//  RssReader
+//
+//  Created by 若江照仁 on 2021/01/07.
+//
+
+import Foundation
+import FirebaseUI
+
+class CommonRouter {
+    /// 認証画面へ
+    /// - Parameter view: FUIAuthDelegateが必要
+    /// SplashView = Transitioner + FUIAuthDelegate
+    class func toAuth(view: SplashViewProtocol) {
+        let authUI = FUIAuth.defaultAuthUI()!
+        authUI.delegate = view
+        authUI.providers = [
+            FUIGoogleAuth(),
+            FUIOAuth.twitterAuthProvider(),
+            FUIEmailAuth()
+        ]
+        let authViewController = authUI.authViewController()
+        authViewController.modalPresentationStyle = .fullScreen
+        view.present(authViewController, animated: true, completion: nil)
+    }
+    
+    class func toArticleListView(view: Transitioner) {
+        let storyboard = UIStoryboard(name: "ArticleList", bundle: nil)
+        let articleListViewController = storyboard.instantiateViewController(identifier: "ArticleListViewController") as! ArticleListViewController
+        articleListViewController.navigationItem.largeTitleDisplayMode = .automatic
+        articleListViewController.navigationItem.title = "記事一覧"
+        
+        let nav = UINavigationController(rootViewController: articleListViewController)
+        nav.modalPresentationStyle = .fullScreen
+        view.present(nav,animated: true, completion: nil)
+    }
+    
+    class func toSelectRssFeedView(view: Transitioner) {
+        let storyboard = UIStoryboard(name: "SelectRssFeed", bundle: nil)
+        let selectRssFeedViewController = storyboard.instantiateViewController(identifier: "SelectRssFeedViewController") as! SelectRssFeedViewController
+        selectRssFeedViewController.navigationItem.largeTitleDisplayMode = .automatic
+        selectRssFeedViewController.navigationItem.title = "記事の選択"
+        selectRssFeedViewController.inject(selectRssFeedRouter: SelectRssFeedRouter(view: selectRssFeedViewController))
+        let nav = UINavigationController(rootViewController: selectRssFeedViewController)
+        nav.modalPresentationStyle = .fullScreen
+        view.present(nav,animated: true, completion: nil)
+    }
+}
